chore(dynamic-routes): consolidated the dynamic-routes logic, and updated docs
